+++ internal/server/udp_pktinfo_stub.go
@@ -1,0 +1,9 @@
+//go:build !linux
+
+package server
+
+import "net"
+
+func wrapUDPPacketInfoImpl(conn *net.UDPConn) UDPConn {
+	return conn
+}